+++ .storybook/preview.ts
@@ -1,0 +1,28 @@
+import type { Preview } from '@storybook/react';
+import { withThemeFromJSXProvider } from '@storybook/addon-themes';
+import { CssBaseline, ThemeProvider } from '@mui/material';
+
+import { voltaireTheme } from '../src/theme';
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  decorators: [
+    withThemeFromJSXProvider({
+      themes: {
+        voltaire: voltaireTheme,
+      },
+      defaultTheme: 'voltaire',
+      Provider: ThemeProvider,
+      GlobalStyles: CssBaseline,
+    }),
+  ],
+};
+
+export default preview;